backend/downloadworker/requirements.txt
cassandra-driver
pika
redis
requests
requests_oauthlib